Make kubejob usable as a library
